+++ test/UT/AssetTest.hs
@@ -93,7 +93,7 @@ loanTests =
                  0.06
                  24
                  P.Current
-      asofDate = (L.toDate "20200615")
+      asofDate = L.toDate "20200615"
       loan1Cf = P.calcCashflow loan1 asofDate
       loan2Cf = P.projCashflow loan1 asofDate []
     in 
@@ -102,9 +102,13 @@ loanTests =
            assertEqual "project period"
              24 
              (CF.sizeCashFlowFrame loan1Cf)
+       ,testCase "First cashflow" $
+           assertEqual ""
+            (Just (CF.LoanFlow (L.toDate "20210201") 120 0 0.61 0 0 0 0 0.06))
+            (CF.cfAt loan1Cf 0)
        ,testCase "Last Principal Amount" $
            assertEqual ""
-            (Just (CF.LoanFlow (L.toDate "20220601") 0 120 0.61 0 0 0 0 0.06))
+            (Just (CF.LoanFlow (L.toDate "20230101") 0 120 0.61 0 0 0 0 0.06))
             (CF.cfAt loan1Cf 23)
        ,testCase "calcCashflow == projCashflow when assump = []" $
            assertEqual ""
@@ -279,22 +283,22 @@ armTest =
     arm1 = ACM.AdjustRateMortgage
             (P.MortgageOriginalInfo 
               240 
-              (Floater2 SOFR3M 0.01 0.03 (EveryNMonth (L.toDate "20231101") 2))
+              (Floater2 SOFR3M 0.01 0.03 (EveryNMonth (L.toDate "20240801") 2))
               30
               Monthly
               (L.toDate "20230501")
               P.Level)
-            (ARM 6 (Just 0.015) (Just 0.01) (Just 0.09) (Just 0.02) )  
+            (ARM 12 (Just 0.015) (Just 0.01) (Just 0.09) (Just 0.02) )  
             240 0.08 19 
             Nothing 
             P.Current
     assump1 = [A.InterestRateCurve 
                 SOFR3M
-                (IRateCurve [TsPoint (L.toDate "20230501") 0.03 
-                            ,TsPoint (L.toDate "20230901") 0.06
-                            ,TsPoint (L.toDate "20231215") 0.07
-                            ,TsPoint (L.toDate "20240315") 0.10
-                            ,TsPoint (L.toDate "20241015") 0.12
+                (IRateCurve [TsPoint (L.toDate "20240501") 0.05 
+                            ,TsPoint (L.toDate "20240901") 0.065
+                            ,TsPoint (L.toDate "20241215") 0.07
+                            ,TsPoint (L.toDate "20250315") 0.10
+                            ,TsPoint (L.toDate "20251001") 0.12
                             ])
                 ]
     arm1_cf = P.projCashflow arm1 (L.toDate "20230601") assump1
@@ -304,29 +308,37 @@ armTest =
         assertEqual "should be 19"
         19
         (CF.sizeCashFlowFrame arm1_cf)
-      ,testCase "ARM case 1" $
-        assertEqual "first rate"
-        (Just (CF.MortgageFlow (L.toDate "20230601") 227.66 12.34 0.6 0 0 0 0 0.03 Nothing ))
+      ,testCase "ARM case 1/ first cash" $
+        assertEqual "first cash row"
+        (Just (CF.MortgageFlow (L.toDate "20240501") 227.66 12.34 0.6 0 0 0 0 0.03 Nothing ))
         (CF.cfAt arm1_cf 0)
-      ,testCase "ARM case 1" $
+      ,testCase "ARM case 1/ frist reset" $
         assertEqual "first rate"
-        (Just (CF.MortgageFlow (L.toDate "20230701") 215.27 12.39 0.56 0 0 0 0 0.03 Nothing ))
+        (Just (CF.MortgageFlow (L.toDate "20240601") 215.41 12.25 0.85 0 0 0 0 0.045 Nothing ))
         (CF.cfAt arm1_cf 1)
-      ,testCase "ARM case 1" $
+      ,testCase "ARM case 1/periodic reset " $
+        assertEqual "first rate"
+        (Just (CF.MortgageFlow (L.toDate "20240801") 190.85 12.26 0.93 0 0 0 0 0.055 Nothing ))
+        (CF.cfAt arm1_cf 3)
+      ,testCase "ARM case 1/remains same before next reset" $
         assertEqual "period before first reset"
-        (Just (CF.MortgageFlow (L.toDate "20231101") 165.44 12.50 0.44 0 0 0 0 0.03 Nothing ))
-        (CF.cfAt arm1_cf 5)
-      ,testCase "ARM case 1" $
-        assertEqual "first reset with cap"
-        (Just (CF.MortgageFlow (L.toDate "20231201") 153 12.44 0.62 0 0 0 0 0.045 Nothing ))
-        (CF.cfAt arm1_cf 6)
+        (Just (CF.MortgageFlow (L.toDate "20240901") 178.53 12.32 0.87 0 0 0 0 0.055 Nothing ))
+        (CF.cfAt arm1_cf 4)
       ,testCase "ARM case 1" $
         assertEqual "reset with periodic cap"
-        (Just (CF.MortgageFlow (L.toDate "20240101") 140.57 12.43 0.7 0 0 0 0 0.055 Nothing ))
+        (Just (CF.MortgageFlow (L.toDate "20241201") 141.47 12.38 0.96 0 0 0 0 0.075 Nothing ))
         (CF.cfAt arm1_cf 7)
       ,testCase "ARM case 1" $
-        assertEqual "life Cap"
-        (Just (CF.MortgageFlow (L.toDate "20241201") 0 13.2 0.09 0 0 0 0 0.09 Nothing ))
-        (CF.cfAt arm1_cf 18)
+        assertEqual "Period 9"
+        (Just (CF.MortgageFlow (L.toDate "20250101") 129.01 12.46 0.88 0 0 0 0 0.075 Nothing ))
+        (CF.cfAt arm1_cf 8)
+      ,testCase "ARM case 1" $
+        assertEqual "Period 10"
+        (Just (CF.MortgageFlow (L.toDate "20250201") 116.49 12.52 0.85 0 0 0 0 0.08 Nothing ))
+        (CF.cfAt arm1_cf 9)
+      ,testCase "ARM case 1" $
+        assertEqual "life cap"
+        (Just (CF.MortgageFlow (L.toDate "20250401") 91.24 12.65 0.77 0 0 0 0 0.09 Nothing ))
+        (CF.cfAt arm1_cf 11)
 
     ]